+++ tagmanager/src/tm_source_file.h
@@ -57,7 +57,7 @@ const gchar *tm_source_file_get_lang_name(gint lang);
 
 gint tm_source_file_get_named_lang(const gchar *name);
 
-gboolean tm_source_file_parse(TMSourceFile *source_file, guchar* text_buf, gint buf_size,
+gboolean tm_source_file_parse(TMSourceFile *source_file, guchar* text_buf, gsize buf_size,
 	gboolean use_buffer);
 
 #endif /* GEANY_PRIVATE */
